untangle zoci with separate load package
