+++ src/internal/packager2/layout/layout.go
@@ -5,13 +5,7 @@
 package layout
 
 import (
-	"context"
 	"path/filepath"
-
-	goyaml "github.com/goccy/go-yaml"
-
-	"github.com/zarf-dev/zarf/src/api/v1alpha1"
-	"github.com/zarf-dev/zarf/src/pkg/logger"
 )
 
 // Constants used in the default package layout.
@@ -51,17 +45,3 @@ const (
 	DataComponentDir      ComponentDir = "data"
 	ValuesComponentDir    ComponentDir = "values"
 )
-
-// ParseZarfPackage parses the yaml passed as a byte slice and applies potential schema migrations.
-func ParseZarfPackage(ctx context.Context, b []byte) (v1alpha1.ZarfPackage, error) {
-	var pkg v1alpha1.ZarfPackage
-	err := goyaml.Unmarshal(b, &pkg)
-	if err != nil {
-		return v1alpha1.ZarfPackage{}, err
-	}
-	pkg, warnings := migrateDeprecated(pkg)
-	for _, warning := range warnings {
-		logger.From(ctx).Warn(warning)
-	}
-	return pkg, nil
-}
