+++ src/internal/packager2/load/deprecated.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2021-Present The Zarf Authors
 
-package layout
+package load
 
 import (
 	"fmt"
